+++ src/main/java/dev/abarmin/aml/dashboard/block/social/SocialNetworksBlockProps.java
@@ -1,7 +1,6 @@
 package dev.abarmin.aml.dashboard.block.social;
 
 import dev.abarmin.aml.dashboard.domain.BlockProps;
-import dev.abarmin.aml.dashboard.domain.BlockType;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,8 +19,4 @@ public class SocialNetworksBlockProps implements BlockProps {
   @Builder.Default
   private List<SocialNetworkLink> links = new ArrayList<>();
 
-  @Override
-  public BlockType type() {
-    return BlockType.SOCIAL_NETWORKS_BLOCK;
-  }
 }