+++ src/main/java/dev/abarmin/aml/dashboard/domain/BlockProps.java
@@ -3,8 +3,8 @@ package dev.abarmin.aml.dashboard.domain;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import dev.abarmin.aml.dashboard.block.avatar.AvatarBlockProps;
+import dev.abarmin.aml.dashboard.block.button.LinkButtonBlockProps;
 import dev.abarmin.aml.dashboard.block.header.HeaderBlockProps;
-import dev.abarmin.aml.dashboard.block.link.LinkButtonBlockProps;
 import dev.abarmin.aml.dashboard.block.social.SocialNetworksBlockProps;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
@@ -15,5 +15,5 @@ import dev.abarmin.aml.dashboard.block.social.SocialNetworksBlockProps;
   @JsonSubTypes.Type(value = SocialNetworksBlockProps.class, name = "SOCIAL_NETWORKS_BLOCK")
 })
 public interface BlockProps {
-  BlockType type();
+  
 }